USERGRID-2458: changes to make compatible with 64 bit (mostly changing int to NSUInteger

+++ source/Classes/AppServices/ApigeeClientResponse.h
@@ -63,7 +63,7 @@ enum
 
 - (id)initWithDataClient:(ApigeeDataClient*)theDataClient;
 
-- (int)entityCount;
+- (NSUInteger)entityCount;
 - (NSArray*)entities;
 - (ApigeeEntity*)firstEntity;
 - (ApigeeEntity*)lastEntity;
